Store physics parameters as f32
`speed` and `fall_acceleration` are used as `f32`, so might as well
store them that way rather than converting from `i32`.

+++ rust/src/player.rs
@@ -5,9 +5,9 @@ use godot::prelude::*;
 #[class(base=CharacterBody3D)]
 pub struct Player {
     #[export]
-    speed: i32, // Speed in m/s
+    speed: f32, // Speed in m/s
     #[export]
-    fall_acceleration: i32, // Acceleration in m/s^2
+    fall_acceleration: f32, // Acceleration in m/s^2
     target_velocity: Vector3,
 
     base: Base<CharacterBody3D>,
@@ -17,8 +17,8 @@ pub struct Player {
 impl ICharacterBody3D for Player {
     fn init(base: Base<CharacterBody3D>) -> Self {
         Self {
-            speed: 14,
-            fall_acceleration: 75,
+            speed: 14.0,
+            fall_acceleration: 75.0,
             target_velocity: Vector3::ZERO,
             base,
         }
@@ -38,13 +38,13 @@ impl ICharacterBody3D for Player {
             pivot.set_basis(Basis::new_looking_at(direction, Vector3::UP, false));
         }
 
-        self.target_velocity.x = direction.x * self.speed as f32;
-        self.target_velocity.z = direction.z * self.speed as f32;
+        self.target_velocity.x = direction.x * self.speed;
+        self.target_velocity.z = direction.z * self.speed;
 
         // Note to self: Not in the tutorial (yet), but shouldn't y velocity
         // reset at some point? (Maybe handled by jumping?)
         if !self.base().is_on_floor() {
-            self.target_velocity.y -= delta as f32 * self.fall_acceleration as f32;
+            self.target_velocity.y -= delta as f32 * self.fall_acceleration;
         }
 
         let target_velocity = self.target_velocity;
